# api/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ResearchFlow API",
    "version": "1.0.0",
    "description": "AI-driven academic research workflow platform API"
  },
  "servers": [
    { "url": "http://localhost:3001", "description": "Development" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/projects": {
      "get": {
        "tags": ["Projects"],
        "summary": "List all projects",
        "responses": { "200": { "description": "Project list" } }
      },
      "post": {
        "tags": ["Projects"],
        "summary": "Create a new project",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["topic"],
                "properties": {
                  "topic": { "type": "string" },
                  "title": { "type": "string" },
                  "description": { "type": "string" },
                  "language": { "type": "string", "enum": ["en", "zh"] }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Project created" } }
      }
    },
    "/api/projects/{id}": {
      "get": {
        "tags": ["Projects"],
        "summary": "Get project details with papers, sections, references",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Project details" }, "404": { "description": "Not found" } }
      }
    },
    "/api/search": {
      "post": {
        "tags": ["Search"],
        "summary": "Search academic papers across multiple sources",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["query"],
                "properties": {
                  "query": { "type": "string" },
                  "sources": { "type": "array", "items": { "type": "string" } },
                  "maxResults": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Search results" } }
      }
    },
    "/api/generate/step/keywords": {
      "post": {
        "tags": ["Generation"],
        "summary": "Step 1: Extract keywords from topic",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["projectId", "topic"],
                "properties": {
                  "projectId": { "type": "string" },
                  "topic": { "type": "string" },
                  "language": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Keywords extracted" } }
      }
    },
    "/api/generate/step/write": {
      "post": {
        "tags": ["Generation"],
        "summary": "Step 5: Generate paper content",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["projectId"],
                "properties": {
                  "projectId": { "type": "string" },
                  "citationFormat": { "type": "string", "enum": ["bibtex", "gbt", "apa", "ieee"] },
                  "language": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Paper generated" } }
      }
    },
...
  }
}